Allowing management url scheme to be set via discovery metadata
Currently, the DefaultServiceInstanceConverter is accessing the metadata to override the address, port and path for management access. There is not way to override the scheme, this is purely set based on getServiceUrl's scheme.

With this change, the scheme is overridable too. Without setting management.scheme to override it, the previous behavior is not breaking.